# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Course System API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Swagger-doc Course System API"
        }
    ],
    "paths": {
...
        "/api/lecturer/create": {
            "post": {
                "tags": [
                    "Lecturer"
                ],
                "summary": "建立新講師",
                "description": "暫無描述。",
                "operationId": "f542ce54cc0b6adcf8718c87a6c46045",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "num": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "department": {
                                        "type": "string"
                                    },
                                    "fulltime": {
                                        "type": "string"
                                    },
                                    "rank": {
                                        "type": "string"
                                    },
                                    "specialty": {
                                        "type": "string"
                                    },
                                    "remark": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "num": "P0006",
                                    "name": "郭台鳴",
                                    "department": "電子工程系",
                                    "fulltime": "專任",
                                    "rank": "教授",
                                    "specialty": "最會賺錢的。",
                                    "remark": "無"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "請求成功。",
                        "content": {
                            "application\\json": {
                                "schema": {
                                    "example": {
                                        "Status": "Success",
                                        "Data": {
                                            "id": 6,
                                            "num": "P0006",
                                            "name": "郭台鳴",
                                            "department": "電子工程系",
                                            "fulltime": "專任",
                                            "rank": "教授",
                                            "specialty": "最會賺錢的。",
                                            "remark": "無"
                                        },
                                        "TimeStamp": "2023-09-07T03:34:32.580848+08:00 CST"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lecturer/show/{id}": {
            "get": {
                "tags": [
                    "Lecturer"
                ],
                "summary": "取得講師詳細資訊",
                "description": "暫無描述。",
                "operationId": "bf509199fd4cea06b3a0d8269c32cd62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "4"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功。",
                        "content": {
                            "application\\json": {
                                "schema": {
                                    "example": {
                                        "Status": "Success",
                                        "Data": {
                                            "id": 4,
                                            "num": "P0004",
                                            "name": "陳小二",
                                            "department": "餐飲管理系",
                                            "fulltime": "兼任",
                                            "rank": "講師",
                                            "specialty": "XXXXXXX",
                                            "remark": "無"
                                        },
                                        "TimeStamp": "2023-09-07T03:53:18.837648+08:00 CST"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lecturer/update/{id}": {
            "put": {
                "tags": [
                    "Lecturer"
                ],
                "summary": "更新講師資訊",
                "description": "暫無描述。",
                "operationId": "639e0ac6d1b958069c0553ac957b52bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "5"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "department": {
                                        "type": "string"
                                    },
                                    "fulltime": {
                                        "type": "string"
                                    },
                                    "rank": {
                                        "type": "string"
                                    },
                                    "specialty": {
                                        "type": "string"
                                    },
                                    "remark": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "黃體操",
                                    "department": "運動休閒系",
                                    "fulltime": "兼任",
                                    "rank": "講師",
                                    "specialty": "體操、籃球、排球",
                                    "remark": "無"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "請求成功。",
                        "content": {
                            "application\\json": {
                                "schema": {
                                    "example": {
                                        "Status": "Success",
                                        "Data": {
                                            "id": 5,
                                            "num": "P0005",
                                            "name": "黃體操",
                                            "department": "運動休閒系",
                                            "fulltime": "兼任",
                                            "rank": "講師",
                                            "specialty": "體操、籃球、排球",
                                            "remark": "限多媒體學院"
                                        },
                                        "TimeStamp": "2023-09-07T03:50:16.208984+08:00 CST"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lecturer/delete/{id}": {
            "delete": {
                "tags": [
                    "Lecturer"
                ],
                "summary": "刪除講師",
                "description": "暫無描述。",
                "operationId": "8c64f311a4cb23b0db14e0d9cf2e83fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功。",
                        "content": {
                            "application\\json": {
                                "schema": {
                                    "example": {
                                        "Status": "Success",
                                        "Data": "陳媒體講師已刪除。",
                                        "TimeStamp": "2023-09-07T03:48:57.662615+08:00 CST"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}